refactor[cm]: Move cohesion test data classes to CodeCohesionFactory
